move event.go, fixup AdmissionEvent
